renewed deprecated sections

+++ application/models.py
@@ -25,7 +25,7 @@ class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(64), unique=True)
     password = db.Column(db.String(64))
-    active = db.Column(db.Integer(1))
+    active = db.Column(db.Integer)
 
     def __init__(self, username, password):
         self.username = username
